happy-path script compilation test passing
