+++ src/script/script.rs
@@ -1,5 +1,5 @@
 use crate::fx::{consume_tick, dissolve, ping_pong, repeating};
-use crate::script::parser::Expr;
+use crate::script::parser::{parse_expr, Expr};
 use crate::{Duration, Effect, EffectTimer, Motion};
 use ratatui::layout::{Margin, Rect};
 use ratatui::style::{Color, Style};
@@ -14,9 +14,8 @@ struct EffectCompiler {
 
 
 
-#[derive(Default)]
 pub struct ScriptContext {
-    deserializers: Vec<EffectCompiler>,
+    compilers: Vec<EffectCompiler>,
 }
 
 pub struct ScriptEnv {
@@ -56,7 +55,9 @@ impl EffectCompiler {
 
 impl ScriptContext {
     pub fn new() -> Self {
-        Self::default()
+        register_default_compilers(Self {
+            compilers: Vec::new(),
+        })
     }
 
     pub fn execute(
@@ -64,7 +65,8 @@ impl ScriptContext {
         env: &mut ScriptEnv,
         input: &str,
     ) -> Option<Effect> {
-        unimplemented!("parse input")
+        parse_expr(input)
+            .and_then(|expr| self.compile(env, expr))
     }
 
     fn compile(
@@ -73,7 +75,7 @@ impl ScriptContext {
         input: Expr
     ) -> Option<Effect> {
         match input {
-            Expr::Fx { name, parameters } => self.deserializers
+            Expr::Fx { name, parameters } => self.compilers
                 .iter()
                 .find(|d| d.name == name)
                 .and_then(|d| {
@@ -92,7 +94,7 @@ impl ScriptContext {
         compiler: impl Fn(&mut InputArgs) -> Option<Effect> + 'static
     ) -> Self {
         let mut this = self;
-        this.deserializers.push(EffectCompiler::new(name, compiler));
+        this.compilers.push(EffectCompiler::new(name, compiler));
         this
     }
 }
@@ -173,222 +175,25 @@ mod compilers {
 
 #[cfg(test)]
 mod tests {
+    use crate::Interpolation::QuadOut;
+    use crate::Shader;
     use super::*;
-    use crate::Interpolation;
-    use std::collections::VecDeque;
-
-    fn empty_env() -> BTreeMap<&'static str, Box<dyn Any>> {
-        BTreeMap::new()
-    }
 
     #[test]
-    fn test_duration_parsing() {
-        let binding = empty_env();
-        let mut args = InputArgs::new(
-            vec![
-                Expr::Duration(Duration::from_millis(500)),
-                Expr::U32(1000),
-            ].into(),
-            &binding
-        );
+    fn happy_path_no_bound_vars() {
+        let input = r#"fx::sweep_in(
+                Motion::LeftToRight,
+                10,
+                0,
+                Color::from_u32(0x1d2021),
+                (1000, QuadOut)
+            )"#;
 
-        assert_eq!(args.duration(), Some(Duration::from_millis(500)));
-        assert_eq!(args.duration(), Some(Duration::from_millis(1000)));
-        assert_eq!(args.duration(), None);
-    }
-
-    #[test]
-    fn test_effect_timer_parsing() {
-        let binding = empty_env();
-        let mut args = InputArgs::new(
-            vec![
-                Expr::Timer(EffectTimer::from_ms(500, Interpolation::Linear)),
-                Expr::U32(1000),
-            ].into(),
-            &binding
-        );
-
-        assert_eq!(args.effect_timer(), Some(EffectTimer::from_ms(500, Interpolation::Linear)));
-        assert_eq!(args.effect_timer(), Some(EffectTimer::from_ms(1000, Interpolation::Linear)));
-        assert_eq!(args.effect_timer(), None);
-    }
-
-    #[test]
-    fn test_numeric_parsing() {
-        let binding = empty_env();
-        let mut args = InputArgs::new(
-            vec![
-                Expr::U32(42),
-                Expr::F32(3.14),
-            ].into(),
-            &binding
-        );
-
-        assert_eq!(args.read_u32(), Some(42));
-        assert_eq!(args.read_f32(), Some(3.14));
-        assert_eq!(args.read_u32(), None);
-    }
-
-    #[test]
-    fn test_string_parsing() {
-        let binding = empty_env();
-        let mut args = InputArgs::new(
-            vec![
-                Expr::String("hello".to_string()),
-                Expr::U32(42), // Wrong type
-                Expr::String("world".to_string()),
-            ].into(),
-            &binding
-        );
-
-        assert_eq!(args.string(), Some("hello".to_string()));
-        assert_eq!(args.string(), None); // U32 should be skipped
-        assert_eq!(args.string(), Some("world".to_string()));
-    }
-
-    #[test]
-    fn test_color_parsing() {
-        let binding = empty_env();
-        let mut args = InputArgs::new(
-            vec![
-                Expr::Color(Color::Red),
-                Expr::Color(Color::Blue),
-            ].into(),
-            &binding
-        );
-
-        assert_eq!(args.color(), Some(Color::Red));
-        assert_eq!(args.color(), Some(Color::Blue));
-        assert_eq!(args.color(), None);
-    }
-
-    #[test]
-    fn test_style_parsing() {
-        let style = Style::default().fg(Color::Red);
-        let binding = empty_env();
-        let mut args = InputArgs::new(
-            vec![
-                Expr::Style(style),
-            ].into(),
-            &binding
-        );
-
-        assert_eq!(args.style(), Some(style));
-        assert_eq!(args.style(), None);
-    }
-
-    #[test]
-    fn test_motion_parsing() {
-        let binding = empty_env();
-        let mut args = InputArgs::new(
-            vec![
-                Expr::Motion(Motion::LeftToRight),
-                Expr::Motion(Motion::UpToDown),
-            ].into(),
-            &binding
-        );
-
-        assert_eq!(args.motion(), Some(Motion::LeftToRight));
-        assert_eq!(args.motion(), Some(Motion::UpToDown));
-        assert_eq!(args.motion(), None);
-    }
-
-    #[test]
-    fn test_margin_parsing() {
-        let margin = Margin::new(10, 20);
-        let binding = empty_env();
-        let mut args = InputArgs::new(
-            vec![
-                Expr::Margin(margin),
-            ].into(),
-            &binding
-        );
-
-        assert_eq!(args.margin(), Some(margin));
-        assert_eq!(args.margin(), None);
-    }
-
-    #[test]
-    fn test_rect_parsing() {
-        let rect = Rect::new(0, 0, 100, 100);
-        let binding = empty_env();
-        let mut args = InputArgs::new(
-            vec![
-                Expr::Rect(rect),
-            ].into(),
-            &binding
-        );
-
-        assert_eq!(args.rect(), Some(rect));
-        assert_eq!(args.rect(), None);
-    }
-
-    #[test]
-    fn test_effect_parsing() {
-        let binding = empty_env();
-        let mut args = InputArgs::new(
-            vec![
-                Expr::Fx {
-                    name: "test".to_string(),
-                    parameters: vec![Expr::U32(500)]
-                },
-            ].into(),
-            &binding
-        );
-
-        // Currently returns None as noted in the TODO
-        assert!(args.effect().is_none());
-    }
-
-    #[test]
-    fn test_mixed_arguments() {
-        let binding = empty_env();
-        let mut args = InputArgs::new(
-            vec![
-                Expr::U32(500),
-                Expr::Motion(Motion::LeftToRight),
-                Expr::Color(Color::Blue),
-                Expr::Timer(EffectTimer::from_ms(1000, Interpolation::Linear)),
-            ].into(),
-            &binding
-        );
-
-        assert_eq!(args.read_u32(), Some(500));
-        assert_eq!(args.motion(), Some(Motion::LeftToRight));
-        assert_eq!(args.color(), Some(Color::Blue));
-        assert_eq!(args.effect_timer(), Some(EffectTimer::from_ms(1000, Interpolation::Linear)));
-        assert_eq!(args.read_u32(), None);
-    }
-
-    #[test]
-    fn test_u16_conversion() {
-        let binding = empty_env();
-        let mut args = InputArgs::new(
-            vec![
-                Expr::U32(65535), // Max u16
-                Expr::U32(65536), // Too large for u16
-            ].into(),
-            &binding
-        );
-
-        assert_eq!(args.read_u16(), Some(65535));
-        assert_eq!(args.read_u16(), Some(0)); // Truncated
-    }
-
-    #[test]
-    fn test_empty_args() {
-        let binding = empty_env();
-        let mut args = InputArgs::new(VecDeque::new(), &binding);
-
-        assert_eq!(args.duration(), None);
-        assert_eq!(args.effect_timer(), None);
-        assert_eq!(args.read_u32(), None);
-        assert_eq!(args.read_f32(), None);
-        assert_eq!(args.string(), None);
-        assert_eq!(args.color(), None);
-        assert_eq!(args.style(), None);
-        assert_eq!(args.motion(), None);
-        assert_eq!(args.margin(), None);
-        assert_eq!(args.rect(), None);
+        let ctx = ScriptContext::new();
+        let effect = ctx.execute(&mut ScriptEnv::new(), input)
+            .expect("effect to be compiled");
+        
+        assert_eq!(effect.name(), "sweep_in");
+        assert_eq!(effect.timer(), Some(EffectTimer::from_ms(1000, QuadOut)));
     }
 }
